+++ R/plot.R
@@ -35,14 +35,16 @@ autoplot.tbl_ts <- function(object, .vars = NULL, ...){
   nk <- n_keys(object)
   
   if(quo_is_null(quo_vars)){
-    if(is_empty(measured_vars(object))){
-      abort("There are no variables to plot.")
+    mv <- measured_vars(object)
+    pos <- which(vapply(object[mv], is.numeric, logical(1L)))
+    if(is_empty(pos)) {
+      abort("Could not automatically identify an appropriate plot variable, please specify the variable to plot.")
     }
     inform(sprintf(
       "Plot variable not specified, automatically selected `.vars = %s`",
-      measured_vars(object)[1]
+      mv[pos[1]]
     ))
-    y <- sym(measured_vars(object)[1])
+    y <- sym(mv[pos[1]])
     .vars <- as_quosures(list(y), env = empty_env())
   }
   else if(possibly(compose(is_quosures, eval_tidy), FALSE)(.vars)){
@@ -100,14 +102,16 @@ autolayer.tbl_ts <- function(object, .vars = NULL, ...){
   nk <- n_keys(object)
   
   if(quo_is_null(quo_vars)){
-    if(is_empty(measured_vars(object))){
-      abort("There are no variables to plot.")
+    mv <- measured_vars(object)
+    pos <- which(vapply(object[mv], is.numeric, logical(1L)))
+    if(is_empty(pos)) {
+      abort("Could not automatically identify an appropriate plot variable, please specify the variable to plot.")
     }
     inform(sprintf(
       "Plot variable not specified, automatically selected `.vars = %s`",
-      measured_vars(object)[1]
+      mv[pos[1]]
     ))
-    y <- sym(measured_vars(object)[1])
+    y <- sym(mv[pos[1]])
     .vars <- as_quosures(list(y), env = empty_env())
   }
   else if(possibly(compose(is_quosures, eval_tidy), FALSE)(.vars)){
@@ -415,6 +419,7 @@ build_fbl_layer <- function(object, data = NULL, level = c(80, 95),
 #' @param object A dable.
 #' @param .vars The column of the dable used to plot. By default, this will be the response variable of the decomposition.
 #' @param scale_bars If `TRUE`, each facet will include a scale bar which represents the same units across each facet.
+#' @param level If the decomposition contains distributions, which levels should be used to display intervals?
 #' @inheritParams autoplot.tbl_ts
 #' 
 #' @examples 
@@ -429,7 +434,8 @@ build_fbl_layer <- function(object, data = NULL, level = c(80, 95),
 #' 
 #' @importFrom ggplot2 ggplot geom_line geom_rect facet_grid vars ylab labs
 #' @export
-autoplot.dcmp_ts <- function(object, .vars = NULL, scale_bars = TRUE, ...){
+autoplot.dcmp_ts <- function(object, .vars = NULL, scale_bars = TRUE, 
+                             level = c(80, 95), ...){
   method <- object%@%"method"
   idx <- index(object)
   keys <- key(object)
@@ -445,16 +451,35 @@ autoplot.dcmp_ts <- function(object, .vars = NULL, scale_bars = TRUE, ...){
   }
   object <- as_tsibble(object) %>% 
     transmute(!!.vars, !!!syms(all.vars(dcmp))) %>% 
-    gather(".var", ".val", !!!syms(measured_vars(.)), factor_key = TRUE)
+    pivot_longer(measured_vars(.), values_to = ".val",
+                 names_to = ".var", names_transform = list(.var = ~ factor(., levels = unique(.))))
   
-  line_aes <- aes(x = !!idx, y = !!sym(".val"))
-  if(n_keys > 1){
-    line_aes$colour <- expr(interaction(!!!keys, sep = "/"))
+  if(has_dist <- inherits(object[[".val"]], "distribution")) { 
+    interval_data <- as_tibble(object)
+    interval_data[paste0(level, "%")] <- lapply(level, hilo, x = interval_data[[".val"]])
+    interval_data <- tidyr::pivot_longer(
+      interval_data, paste0(level, "%"), names_to = NULL, values_to = "hilo"
+    )
+    intvl_aes <- aes(x = !!idx, hilo = !!sym("hilo"))
+    line_aes <- aes(x = !!idx, y = mean(!!sym(".val")))
+    if(n_keys > 1){
+      line_aes$colour <- intvl_aes$fill <- expr(interaction(!!!keys, sep = "/"))
+    }
+    dcmp_geom <- list(
+      distributional::geom_hilo_ribbon(intvl_aes, ..., data = interval_data),
+      geom_line(line_aes, ...)
+    )
+  } else {
+    line_aes <- aes(x = !!idx, y = !!sym(".val"))
+    if(n_keys > 1){
+      line_aes$colour <- expr(interaction(!!!keys, sep = "/"))
+    }
+    dcmp_geom <- geom_line(line_aes, ...)
   }
   
   p <- object %>% 
     ggplot() + 
-    geom_line(line_aes, ...) + 
+    dcmp_geom + 
     facet_grid(vars(!!sym(".var")), scales = "free_y") + 
     ylab(NULL) + 
     labs(
@@ -470,10 +495,12 @@ autoplot.dcmp_ts <- function(object, .vars = NULL, scale_bars = TRUE, ...){
     # Avoid issues with visible bindings
     ymin <- ymax <- center <- diff <- NULL
     
-    range_data <- object %>%
-      as_tibble %>% 
+    min_fn <- if(has_dist) function(x, ...) min(quantile(x, (100-max(level))/200), ...) else min
+    max_fn <- if(has_dist) function(x, ...) max(quantile(x, (100 + max(level))/200), ...) else max
+    
+    range_data <- as_tibble(object) %>%
       group_by(!!sym(".var")) %>% 
-      summarise(ymin = min(!!sym(".val"), na.rm = TRUE), ymax = max(!!sym(".val"), na.rm = TRUE)) %>% 
+      summarise(ymin = min_fn(!!sym(".val"), na.rm = TRUE), ymax = max_fn(!!sym(".val"), na.rm = TRUE)) %>% 
       mutate(
         center = (ymin + ymax) / 2,
         diff = min(ymax - ymin),