+++ NEWS.md
@@ -1,21 +1,49 @@
-# fabletools (development version)
+# fabletools 0.3.0
 
 ## New features
 
+* Added `scenarios()` function for providing multiple scenarios to the 
+  `new_data` argument. This allows different sets of future exogenous regressors
+  to be provided to functions like `forecast()`, `generate()`, and 
+  `interpolate()` (#110).
 * Added `quantile_score()`, which is similar to `percentile_score()` except it
   allows a set of quantile `probs` to be provided (#280).
+* Added distribution support for `autoplot(<dable>)`. If the decomposition 
+  provides distributions for its components, then the uncertainty of the 
+  components will be plotted with interval ribbons.
+* Added block bootstrap option for bootstrapping innovations in `generate()`.
+* Added multiple step ahead fitted values support via `fitted(<mable>, h > 1)`.
+* Added `as_fable(<forecast>)` for converting older `forecast` class objects to
+  `fable` data structures.
+* Added `top_down(method = "forecast_proportion")` for reconciliation using the
+  forecast proportions techniques.
+* Added `middle_out()` forecast reconciliation method.
+* Added directional accuracy measures, including `MDA()`, `MDV()` and `MDPV()`
+  (#273, @davidtedfordholt).
+* Added `fill_gaps(<fable>)`.
 
 ## Improvements
 
 * The `pinball_loss()` and `percentile_score()` accuracy measures are now scaled
   up by 2x for improved meaning. The loss at 50% equals absolute error and the
   average loss equals CRPS (#280).
+* Automatic transformation functions formals are now named after the response
+  variable and not converted to `.x`, preventing conflicts with values named `.x`.
+* `box_cox()` and `inv_box_cox()` are now vectorised over the transformation
+  parameter `lambda`.
+* `RMSSE()` accuracy measure is now included in default `accuracy()` measures.
+* Specifying a different `response` variable in `as_fable()` will no longer
+  error, it now sets the provided `response` value as the distribution's new 
+  response.
+* Minor vctrs support improvements.
 
 ## Bug fixes
 
 * Data lines in fable `autoplot()` are now always grouped by the data's key.
 * Fixed `bottom_up()` aggregation mismatch for redundant leaf nodes (#266).
 * Fixed `min_trace()` reconciliation for degenerate hierarchies (#267).
+* Fixed `select(<mable>)` not keeping required key variables (#297).
+* Fixed `...` not being passed through in `report()`.
 
 # fabletools 0.2.1